+++ src/scancode_required_phrases/__init__.py
@@ -1,0 +1,6 @@
+# -*- coding: utf-8 -*-
+#
+# Copyright (c) nexB Inc. and others. All rights reserved.
+# SPDX-License-Identifier: Apache-2.0
+
+"""Tools for building required phrase datasets from ScanCode license rules."""